# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AMPARO - Sistema de Gestão de Casas de Repouso
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /residents:
    get:
      operationId: listResidents
      tags: [residents]
      summary: List all residents
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: Residents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resident"
    post:
      operationId: createResident
      tags: [residents]
      summary: Create a new resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidentInput"
      responses:
        "201":
          description: Resident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
  /residents/{id}:
    get:
      operationId: getResident
      tags: [residents]
      summary: Get resident by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
        "404":
          description: Not found
    patch:
      operationId: updateResident
      tags: [residents]
      summary: Update resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidentUpdate"
      responses:
        "200":
          description: Resident updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
    delete:
      operationId: deleteResident
      tags: [residents]
      summary: Delete resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /residents/{id}/profile:
    get:
      operationId: getResidentProfile
      tags: [residents]
      summary: Get full resident profile with history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full resident profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentProfile"
  /medications:
    get:
      operationId: listMedications
      tags: [medications]
      summary: List medications
      parameters:
        - name: residentId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [administered, pending, late]
      responses:
        "200":
          description: Medications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [medications]
      summary: Add medication to resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationInput"
      responses:
        "201":
          description: Medication created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
  /medications/{id}:
    patch:
      operationId: updateMedication
      tags: [medications]
      summary: Update medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationUpdate"
      responses:
        "200":
          description: Updated medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
    delete:
      operationId: deleteMedication
      tags: [medications]
      summary: Delete medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /medications/{id}/administer:
    post:
      operationId: administerMedication
      tags: [medications]
      summary: Record medication administration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdministrationInput"
      responses:
        "201":
          description: Administration recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationAdministration"
  /medications/{id}/administrations:
    get:
      operationId: getMedicationAdministrations
      tags: [medications]
      summary: Get administration history for medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Administration history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicationAdministration"
...
components:
  schemas:
...
    Resident:
      type: object
      required: [id, name, dateOfBirth, gender, room, admissionDate, status]
      properties:
        id:
          type: integer
        name:
          type: string
        cpf:
          type: ["string", "null"]
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        room:
          type: string
        admissionDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive]
        diagnoses:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        familyContact:
          type: ["string", "null"]
        familyPhone:
          type: ["string", "null"]
        clinicalNotes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResidentInput:
      type: object
      required: [name, dateOfBirth, gender, room, admissionDate]
      properties:
        name:
          type: string
        cpf:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        room:
          type: string
        admissionDate:
          type: string
          format: date
        diagnoses:
          type: string
        allergies:
          type: string
        familyContact:
          type: string
        familyPhone:
          type: string
        clinicalNotes:
          type: string
    ResidentUpdate:
      type: object
      properties:
        name:
          type: string
        cpf:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        room:
          type: string
        admissionDate:
          type: string
        diagnoses:
          type: string
        allergies:
          type: string
        familyContact:
          type: string
        familyPhone:
          type: string
        clinicalNotes:
          type: string
        status:
          type: string
          enum: [active, inactive]
    ResidentProfile:
      type: object
      required: [resident, medications, activities, appointments, handoverNotes]
      properties:
        resident:
          $ref: "#/components/schemas/Resident"
        medications:
          type: array
          items:
            $ref: "#/components/schemas/Medication"
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        handoverNotes:
          type: array
          items:
            type: string
    Medication:
      type: object
      required: [id, residentId, name, dose, frequency, scheduledTimes, status]
      properties:
        id:
          type: integer
        residentId:
          type: integer
        residentName:
          type: ["string", "null"]
        name:
          type: string
        dose:
          type: string
        frequency:
          type: string
        scheduledTimes:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [administered, pending, late]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MedicationInput:
      type: object
      required: [residentId, name, dose, frequency, scheduledTimes]
      properties:
        residentId:
          type: integer
        name:
          type: string
        dose:
          type: string
        frequency:
          type: string
        scheduledTimes:
          type: array
          items:
            type: string
        notes:
          type: string
    MedicationUpdate:
      type: object
      properties:
        name:
          type: string
        dose:
          type: string
        frequency:
          type: string
        scheduledTimes:
          type: array
          items:
            type: string
        notes:
          type: string
        status:
          type: string
          enum: [administered, pending, late]
    MedicationAdministration:
      type: object
      required: [id, medicationId, administeredAt, administeredBy]
      properties:
        id:
          type: integer
        medicationId:
          type: integer
        administeredAt:
          type: string
          format: date-time
        administeredBy:
          type: string
        notes:
          type: ["string", "null"]
        isLate:
          type: boolean
        lateJustification:
          type: ["string", "null"]
    AdministrationInput:
      type: object
      required: [administeredAt, administeredBy]
      properties:
        administeredAt:
          type: string
          format: date-time
        administeredBy:
          type: string
        notes:
          type: string
        lateJustification:
          type: string
    Activity:
      type: object
      required: [id, residentId, type, date, status]
      properties:
        id:
          type: integer
        residentId:
          type: integer
        residentName:
          type: ["string", "null"]
        type:
          type: string
          enum: [bath, feeding, hygiene, physiotherapy, medication, other]
        date:
          type: string
          format: date
        time:
          type: ["string", "null"]
        status:
          type: string
          enum: [completed, pending, not_done]
        observations:
          type: ["string", "null"]
        shift:
          type: string
          enum: [morning, afternoon, night]
        createdAt:
          type: string
          format: date-time
...
    Appointment:
      type: object
      required: [id, residentId, date, time, specialty, status]
      properties:
        id:
          type: integer
        residentId:
          type: integer
        residentName:
          type: ["string", "null"]
        date:
          type: string
          format: date
        time:
          type: string
        specialty:
          type: string
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled]
        observations:
          type: ["string", "null"]
        cancellationReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...