# lib/api-zod/src/generated/types/residentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AMPARO - Sistema de Gestão de Casas de Repouso
 * OpenAPI spec version: 0.1.0
 */
import type { ResidentInputGender } from './residentInputGender';

export interface ResidentInput {
  name: string;
  cpf?: string;
  dateOfBirth: Date;
  gender: ResidentInputGender;
  room: string;
  admissionDate: Date;
  diagnoses?: string;
  allergies?: string;
  familyContact?: string;
  familyPhone?: string;
  clinicalNotes?: string;
}